core: Rename hcid.h to btd.h
Since we use btd term for daemon APIs it makes sense to use it also
for the header name instead of hcid.
