+++ src/plugin.c
@@ -24,7 +24,7 @@
 #include "btio/btio.h"
 #include "src/plugin.h"
 #include "src/log.h"
-#include "src/hcid.h"
+#include "src/btd.h"
 
 static GSList *plugins = NULL;
